refactor: typed JWT payload, Decimal money, narrower excepts
- Typed JWT: verify_token returns a validated TokenPayload (user_id: UUID
  required) instead of a raw dict; get_current_user_id simplified. A token
  missing user_id now 401s at validation.
- Money as Decimal end-to-end: ShopProduct.price_usd is Decimal, adapters
  parse via Decimal(str(...)); price fetcher no longer round-trips through
  float.
- Narrower exception handling in NBU fetch: catch network/parse errors
  (httpx/Value/Key/Type/Arithmetic) instead of bare Exception, so code bugs
  surface instead of being masked.
- Consistent money quantization: convert() now quantizes the USD path too
  (history no longer shows 16-decimal values).
- Repo-wide ruff modernization (Optional[X] -> X | None, typing -> abc) and
  UP017 disabled to keep the datetime.timezone.utc house style.

74 passed, coverage ~79%. Verified live: token-without-user_id 401,
Decimal prices flow, history values now 4-decimal.

+++ app/services/shop_adapters/base.py
@@ -1,5 +1,6 @@
 from abc import ABC, abstractmethod
 from dataclasses import dataclass
+from decimal import Decimal
 
 
 @dataclass
@@ -8,7 +9,7 @@ class ShopProduct:
     title: str
     description: str
     category: str
-    price_usd: float
+    price_usd: Decimal
 
 
 class BaseShopAdapter(ABC):
